requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
blessed==1.18.1
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.3
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
docopt==0.6.2
future==0.16.0
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.3.1
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
texttable==1.6.4
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==5.3.0
django-storages==1.11.1
boto3==1.18.38
botocore==1.21.38
